\name{ClusteringTree4-package}
\alias{ClusteringTree4-package}
\alias{ClusteringTree4}
\docType{package}
\title{
\packageTitle{ClusteringTree4}
}
\description{
\packageDescription{ClusteringTree4}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ClusteringTree4}
\packageIndices{ClusteringTree4}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ClusteringTree4}

Maintainer: \packageMaintainer{ClusteringTree4}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
